refactor: separate amount input item
